% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{choose_eafdiffplot}
\alias{choose_eafdiffplot}
\alias{choose_eafdiff}
\title{Plot empirical attainment function differences}
\usage{
choose_eafdiffplot(data.left, data.right, intervals = 5,
  maximise = c(FALSE, FALSE),
  title.left = deparse(substitute(data.left)),
  title.right = deparse(substitute(data.right)), ...)

choose_eafdiff(x, left = stop("'left' must be either TRUE or FALSE"))
}
\arguments{
\item{data.left, data.right}{Data frames corresponding to the input data of
left and right sides, respectively. Each data frame has at least three
columns, the third one being the set of each point. See also
\code{\link[=read_datasets]{read_datasets()}}.}

\item{intervals}{(\code{integer(1)}|\code{character()}) \cr The absolute range of the
differences \eqn{[0, 1]} is partitioned into the number of intervals
provided. If an integer is provided, then labels for each interval are
computed automatically. If a character vector is provided, its length is
taken as the number of intervals.}

\item{maximise}{(\code{logical()} | \code{logical(1)}) \cr Whether the objectives must be
maximised instead of minimised. Either a single logical value that applies
to all objectives or a vector of logical values, with one value per
objective.}

\item{title.left, title.right}{Title for left and right panels, respectively.}

\item{...}{Other graphical parameters are passed down to
\code{\link[=eafdiffplot]{eafdiffplot()}}.}

\item{x}{(\code{matrix()}) Matrix of rectangles representing EAF differences
(returned by \code{\link[=eafdiff]{eafdiff()}} with \code{rectangles=TRUE}).}

\item{left}{(\code{logical(1)}) With `left=TRUE' return the rectangles with
positive differences, otherwise return those with negative differences but
differences are converted to positive.}
}
\value{
No return value.
}
\description{
Plot the differences between the empirical attainment functions of two
data sets as a two-panel plot, where the left side shows the values of
the left EAF minus the right EAF and the right side shows the
differences in the other direction.
}
\details{
This function calculates the differences between the EAFs of two
data sets, and plots on the left the differences in favour
of the left data set, and on the right the differences in favour of
the right data set. By default, it also plots the grand best and worst
attainment surfaces, that is, the 0\% and 100\%-attainment surfaces
over all data. This two surfaces delimit the area where differences
may exist. In addition, it also plots the 50\%-attainment surface of
each data set.
}
\examples{

\donttest{
extdata_dir <- system.file(package="eaf", "extdata") 
A1 <- read_datasets(file.path(extdata_dir, "wrots_l100w10_dat"))
A2 <- read_datasets(file.path(extdata_dir, "wrots_l10w100_dat"))
if (interactive()) {
  rectangles <- choose_eafdiffplot(A1, A2, intervals = 5)
} else { # Choose A1
  rectangles <- eafdiff(A1, A2, intervals = 5, rectangles = TRUE)
  rectangles <- choose_eafdiff(rectangles, left = TRUE)
}
reference <- c(max(A1[, 1], A2[, 1]), max(A1[, 2], A2[, 2]))
x <- split.data.frame(A1[,1:2], A1[,3])
whv_A1 <- sapply(split.data.frame(A1[, 1:2], A1[, 3]),
                 whv_rect, rectangles=rectangles, reference=reference)
whv_A2 <- sapply(split.data.frame(A2[, 1:2], A2[, 3]),
                 whv_rect, rectangles=rectangles, reference=reference)
boxplot(list(A1=whv_A1, A2=whv_A2), main = "Weighted hypervolume")
}
}
\seealso{
\code{\link[=read_datasets]{read_datasets()}}, \code{\link[=eafdiffplot]{eafdiffplot()}}
}
\keyword{graphs}
