% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{filter_dominated}
\alias{filter_dominated}
\title{Remove dominated points}
\usage{
filter_dominated(data, maximise = FALSE, keep_weakly = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame of numerical values, where
each row gives the coordinates of a point.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single logical value that applies to all objectives or
a vector of logical values, with one value per objective.}

\item{keep_weakly}{If \code{FALSE}, return \code{FALSE} for any duplicates
of nondominated points.}
}
\value{
a matrix or data.frame with only mutually nondominated points.
}
\description{
Remove dominated points
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
