% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf-package.R
\docType{data}
\name{gcp2x2}
\alias{gcp2x2}
\title{Metaheuristics for solving the Graph Vertex Coloring Problem}
\format{
A data frame with 3133 observations on the following 6 variables.
\describe{
\item{\code{alg}}{a factor with levels \code{SAKempeFI} and \code{TSinN1}}
\item{\code{inst}}{a factor with levels \code{DSJC500.5} and
\code{DSJC500.9}. Instances are taken from the DIMACS repository.}
\item{\code{run}}{a numeric vector indicating the run to
which the observation belong. }
\item{\code{best}}{a numeric vector indicating the best solution in
number of colors found in the corresponding run up to that time.}
\item{\code{time}}{a numeric vector indicating the time since the
beginning of the run for each observation. A rescaling is applied.}
\item{\code{titer}}{a numeric vector indicating iteration number
corresponding to the observations.}
}
}
\source{
\insertRef{ChiarandiniPhD}{eaf} (page 138)
}
\usage{
gcp2x2
}
\description{
Two metaheuristic algorithms, TabuCol (Hertz et al., 1987) and
simulated annealing \citep{JohAraMcGSch1991}, to find a good
approximation of the chromatic number of two random graphs. The data
here has the only goal of providing an example of use of eafplot for
comparing algorithm performance with respect to both time and quality
when modelled as two objectives in trade off.
}
\details{
Each algorithm was run 10 times per graph registering the time and
iteration number at which a new best solution was found. A time limit
corresponding to 500*10^5 total iterations of TabuCol was imposed. The
time was then normalized on a scale from 0 to 1 to make it instance
independent.
}
\examples{
data(gcp2x2)
}
\references{
A. Hertz and D. de Werra. Using Tabu Search Techniques for Graph
Coloring. Computing, 1987, 39(4), 345-351.

\insertAllCited{}
}
\keyword{datasets}
