% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{hv_contributions}
\alias{hv_contributions}
\title{Hypervolume contribution of a set of points}
\usage{
hv_contributions(data, reference, maximise = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame of numerical values, where
each row gives the coordinates of a point.}

\item{reference}{Reference point as a vector of numerical values.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single boolean value that applies to all objectives or
a vector boolean values, with one value per objective.}
}
\value{
A numerical vector
}
\description{
Computes the hypervolume contribution of each point given a set of points
with respect to a given reference point assuming minimization of all
objectives. If the set contains duplicated or dominated points, then each individual
point contributes zero.
}
\examples{

data(SPEA2minstoptimeRichmond)
# The second objective must be maximized
# We calculate the hypervolume contribution of each point of the union of all sets.
hv_contributions(SPEA2minstoptimeRichmond[, 1:2], reference = c(250, 0),
            maximise = c(FALSE, TRUE))

# Duplicated points contributed zero even if nondominated, however,
# filter_dominated removes all duplicates except one. Hence, nondominated
# points will have nonzero contribution.
hv_contributions(filter_dominated(SPEA2minstoptimeRichmond[, 1:2], maximise = c(FALSE, TRUE)),
                 reference = c(250, 0), maximise = c(FALSE, TRUE))

}
\references{
C. M. Fonseca, L. Paquete, and M. López-Ibáñez. An improved dimension-sweep
algorithm for the hypervolume indicator. In IEEE Congress on Evolutionary
Computation, pages 1157-1163, Vancouver, Canada, July 2006.

Nicola Beume, Carlos M. Fonseca, Manuel López-Ibáñez, Luís Paquete, and
J. Vahrenhold. On the complexity of computing the hypervolume indicator.
IEEE Transactions on Evolutionary Computation, 13(5):1075-1082, 2009.
}
\seealso{
\code{\link{hypervolume}}
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
