% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf-package.R
\docType{data}
\name{HybridGA}
\alias{HybridGA}
\title{Results of Hybrid GA on vanzyl and Richmond water networks}
\format{
A list with two data frames, each of them with three columns, as
produced by \code{\link[=read_datasets]{read_datasets()}}.
\describe{
\item{\verb{$vanzyl}}{data frame of results on vanzyl network}
\item{\verb{$richmond}}{data frame of results on Richmond
network. The second column is filled with \code{NA}}
}
}
\source{
\insertRef{LopezIbanezPhD}{eaf}.
}
\usage{
HybridGA
}
\description{
The data has the only goal of providing an example of use of eafplot.
}
\examples{
data(HybridGA)
print(HybridGA$vanzyl)
print(HybridGA$richmond)
}
\keyword{datasets}
