% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{is.nondominated}
\alias{is.nondominated}
\title{Identify nondominated points}
\usage{
is.nondominated(data, maximise = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame of numerical values, where
each row gives the coordinates of a point.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single boolean value that applies to all objectives or
a vector boolean values, with one value per objective.}
}
\value{
A logical vector of the same length as the number of rows of
  \code{data}, where \code{TRUE} means that the point is not dominated by
  any other point.
}
\description{
Identify nondominated points
}
\examples{
path.A1 <- file.path(system.file(package="eaf"),"extdata","ALG_1_dat")
path.A2 <- file.path(system.file(package="eaf"),"extdata","ALG_2_dat")
set <- rbind(read.data.sets(path.A1)[,1:2], read.data.sets(path.A2)[,1:2])

is.nondom <- is.nondominated(set)
cat("There are ", sum(is.nondom), " nondominated points\\n")

ndset <- set[is.nondom,]
ndset <- ndset[order(ndset[,1]),]
plot(set, col = "blue", type = "p", pch = 20)
points(ndset, col = "red", pch = 21)

}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
