% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{is_nondominated}
\alias{is_nondominated}
\title{Identify nondominated points}
\usage{
is_nondominated(data, maximise = FALSE, keep_weakly = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame of numerical values, where
each row gives the coordinates of a point.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single logical value that applies to all objectives or
a vector of logical values, with one value per objective.}

\item{keep_weakly}{If \code{FALSE}, return \code{FALSE} for any duplicates
of nondominated points.}
}
\value{
A logical vector of the same length as the number of rows of
  \code{data}, where \code{TRUE} means that the point is not dominated by
  any other point.
}
\description{
Identify nondominated points
}
\examples{
path_A1 <- file.path(system.file(package="eaf"),"extdata","ALG_1_dat")
path_A2 <- file.path(system.file(package="eaf"),"extdata","ALG_2_dat")
set <- rbind(read.data.sets(path_A1)[,1:2], read.data.sets(path_A2)[,1:2])

is_nondom <- is_nondominated(set)
cat("There are ", sum(is_nondom), " nondominated points\\n")

ndset <- set[is_nondom, ] # or ndset <- filter_dominated(set)
ndset <- ndset[order(ndset[,1]),]
plot(set, col = "blue", type = "p", pch = 20)
points(ndset, col = "red", pch = 21)

}
\seealso{
\code{filter_dominated}
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
