% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{pareto_rank}
\alias{pareto_rank}
\title{Pareto ranking (i.e., nondominated sorting)}
\usage{
pareto_rank(data, maximise = FALSE)
}
\arguments{
\item{data}{Either a matrix or a data frame of numerical values, where
each row gives the coordinates of a point.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single logical value that applies to all objectives or
a vector of logical values, with one value per objective.}
}
\value{
An integer vector of the same length as the number of rows of
  \code{data}, where each value gives the rank of each point.
}
\description{
Pareto ranking (i.e., nondominated sorting)
}
\details{
This function is meant to be used like \code{rank()}, but it
  assigns ranks according to Pareto dominance. Duplicated points are kept on
  the same front.
}
\examples{
path.A1 <- file.path(system.file(package="eaf"),"extdata","ALG_1_dat")
set <- read.data.sets(path.A1)[,1:2]
ranks <- pareto_rank(set)
colors <- colorRampPalette(c("red","yellow","springgreen","royalblue"))(max(ranks))
plot(set, col = colors[ranks], type = "p", pch = 20)

}
\references{
Deb, K., S. Agrawal, A. Pratap, and T. Meyarivan. A fast elitist non-dominated
sorting genetic algorithm for multi-objective optimization: NSGA-II.
IEEE Transactions on Evolutionary Computation, 6(2): 182-197, 2002.

M. T. Jensen. Reducing the run-time complexity of multiobjective
EAs: The NSGA-II and other algorithms. IEEE Transactions on
Evolutionary Computation, 7(5):503–515, 2003.
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
