% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf-package.R
\docType{data}
\name{SPEA2relativeVanzyl}
\alias{SPEA2relativeVanzyl}
\title{Results of SPEA2 with relative time-controlled triggers on Vanzyl's
water network.}
\format{A data frame as produced by \code{\link[=read_datasets]{read_datasets()}}.}
\source{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}. Operational Optimisation of Water Distribution
Networks. PhD thesis, School of Engineering and the Built Environment,
Edinburgh Napier University, UK, 2009.
}
\usage{
SPEA2relativeVanzyl
}
\description{
The data has the only goal of providing an example of use of eafplot.
}
\examples{
data(HybridGA)
data(SPEA2relativeVanzyl)
eafplot(SPEA2relativeVanzyl, percentiles = c(25, 50, 75), 
       xlab = expression(C[E]), ylab = "Total switches", xlim = c(320, 400),
       extra.points = HybridGA$vanzyl, extra.legend = "Hybrid GA")
}
\keyword{datasets}
