% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datasets.R
\name{write_datasets}
\alias{write_datasets}
\title{Write data sets}
\usage{
write_datasets(x, file = "")
}
\arguments{
\item{x}{The data set to write. The last column must be the set number.}

\item{file}{either a character string naming a file or a connection open for
writing. ‘""’ indicates output to the console.}
}
\description{
Write data sets to a file in the same format as \code{\link[=read_datasets]{read_datasets()}}.
}
\examples{
x <- read_datasets(text="1 2\n3 4\n\n5 6\n7 8\n", col_names=c("obj1", "obj2"))
write_datasets(x)

}
\seealso{
\code{\link[utils]{write.table}}, \code{\link[=read_datasets]{read_datasets()}}
}
\keyword{file}
