% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hv.R
\name{filter_dominated}
\alias{filter_dominated}
\title{Remove dominated points}
\usage{
filter_dominated(..., maximise = FALSE)
}
\arguments{
\item{...}{matrices or data frames of numerical values, where
each row gives the coordinates of a point.}

\item{maximise}{Whether the objectives must be maximised instead of
minimised. Either a single boolean value that applies to all objectives or
a vector boolean values, with one value per objective.}
}
\value{
a matrix or data.frame with only mutually nondominated points.
}
\description{
Remove dominated points
}
\author{
Manuel \enc{López-Ibáñez}{Lopez-Ibanez}
}
