% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vorob.R
\name{vorobT}
\alias{vorobT}
\alias{vorobDev}
\alias{symDifPlot}
\alias{symDifPlot2}
\title{Vorob'ev computations}
\usage{
vorobT(x, reference)

vorobDev(x, VE, reference)

symDifPlot(x, VE, threshold, add = FALSE, nlevels = 8, ve.col = "red",
  ve.pch = 4)

symDifPlot2(x, VE, threshold, add = FALSE, nlevels = 8, ve.col = "red",
  ve.pch = 4)
}
\arguments{
\item{x}{Either a matrix of data values, or a data frame, or a list of data
frames of exactly three columns.  The third column is assumed to give the
run identifier, from 1 to the number of runs.}

\item{reference}{Reference point as a vector of numerical values.}

\item{VE, threshold}{Vorob'ev expectation and threshold, e.g., as returned
by \code{\link[eaf]{vorobT}}.}

\item{add}{if \code{FALSE}, a new graph is created}

\item{nlevels}{number of levels in which is divided the range of the
symmetric deviation}

\item{ve.col, }{ve.pch plotting parameters for the Vorob'ev expectation}
}
\value{
\code{vorobT} returns a list with elements \code{threshold} and
  \code{VE}

\code{vorobDev} returns the Vorob'ev deviation.
}
\description{
Compute Vorob'ev threshold, expectation and deviation. Also, displaying the
symmetric deviation function is possible.  The symmetric deviation
function is the probability for a given target in the objective space to
belong to the symmetric difference between the Vorob'ev expectation and a
realization of the (random) attained set.
}
\note{
Are x's supposed to be sets of Pareto optimal points?
}
\examples{
data(CPFs)
res <- vorobT(CPFs, reference = c(2, 200))
print(res$threshold)

## Display Vorob'ev expectation and attainment function
# First style
eafplot(CPFs[,1:2], sets = CPFs[,3], percentiles = c(0, 25, 50, 75, 100, res$threshold),
        main = substitute(paste("Empirical attainment function, ",beta,"* = ", a),
                          list(a = formatC(res$threshold, digits = 2))))

# Second style
eafplot(CPFs[,1:2], sets = CPFs[,3], percentiles = c(0, 20, 40, 60, 80, 100),
        col=gray(seq(0.8, 0.1, length.out = 6)^0.5), type = "area", 
        legend.pos = "bottomleft", extra.points = res$VE, extra.col = "cyan", extra.legend = "VE",
        main = substitute(paste("Empirical attainment function, ",beta,"* = ", a),
                          list(a = formatC(res$threshold, digits = 2))))


# Now print Vorob'ev deviation
VD <- vorobDev(CPFs, res$VE, reference = c(2, 200))
print(VD)

# Now display symmetric deviation function
symDifPlot(CPFs, res$VE, res$threshold, add = FALSE, nlevels = 8)
symDifPlot2(CPFs, res$VE, res$threshold, add = FALSE, nlevels = 8)

}
\author{
Mickael Binois
}
